fix: re-direct to plan by id when saved
